+++ Data/Queue.hs
@@ -6,6 +6,8 @@ module Data.Queue
   , empty
   , next
   , remove
+  , peak
+  , depth
   , (|>)
   ) where
 import           Protolude            hiding (empty, on)
@@ -25,6 +27,12 @@ next = on (Seq.drop 1)
 
 remove :: Eq a => a -> Queue a -> Queue a
 remove el = on (Seq.filter (/= el))
+
+peak :: Queue a -> Maybe a
+peak (Q s) = Seq.lookup 0 s
+
+depth :: Queue a -> Int
+depth (Q s) = Seq.length s
 
 on :: (Seq a -> Seq a) -> Queue a -> Queue a
 on f (Q s) = Q $ f s