+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021.05.14
+### Added
+- More sensible stack update when creating new local stack from broad stack
+- Added missing cookie deletion functionality
+
 ## [0.2.1] - 2021-05-14
 ### Fix
 - Delete user also clears removes the user from local and global stacks